validating up count column with int

+++ score_calculator/simple_comment_score_calculator.py
@@ -12,7 +12,7 @@ class SimpleCommentScoreCalculator(BaseCalculator):
     def score(self, *, comments_data_frame: pd.DataFrame):
         if type(comments_data_frame) is not pd.DataFrame:
             raise InvalidArgumentValue("comments data frame needs a pandas data frame")
-        if comments_data_frame.dtypes["ups"] != numpy.float:
+        if comments_data_frame.dtypes["ups"] != numpy.int:
             raise InvalidArgumentValue("comments data frame contains invalid data")
         comments_data_frame["score"] = (comments_data_frame["ups"] - comments_data_frame["ups"].mean()) / \
                                        comments_data_frame["ups"].std()
