+++ score_calculator/simple_post_score_calculator.py
@@ -14,7 +14,7 @@ class SimplePostScoreCalculator(BaseCalculator):
             raise InvalidArgumentValue("post should be a dictionary")
         if type(comments_data_frame) is not pd.DataFrame:
             raise InvalidArgumentValue("comments data frame needs a pandas data frame")
-        if comments_data_frame.dtypes["ups"] != numpy.float:
+        if comments_data_frame.dtypes["ups"] != numpy.int:
             raise InvalidArgumentValue("comments data frame contains invalid data")
         post["score"] = sum(comments_data_frame["score"] / comments_data_frame["score"].count())
         return post
